+++ bootique-swagger/src/main/java/io/bootique/swagger/factory/SwaggerServiceFactory.java
@@ -21,12 +21,13 @@ package io.bootique.swagger.factory;
 import io.bootique.annotation.BQConfig;
 import io.bootique.annotation.BQConfigProperty;
 import io.bootique.resource.ResourceFactory;
+import io.bootique.swagger.OpenApiCustomizer;
 import io.bootique.swagger.OpenApiModel;
 import io.bootique.swagger.SwaggerService;
+import io.bootique.swagger.customizer.PathSortingCustomizer;
+import io.bootique.swagger.customizer.SchemasSortingCustomizer;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
 
 @BQConfig
 public class SwaggerServiceFactory {
@@ -51,22 +52,41 @@ public class SwaggerServiceFactory {
         this.pretty = pretty;
     }
 
-    public SwaggerService createSwaggerService() {
-        var models = createModels(this.specs);
-        return new SwaggerService(models);
+    public SwaggerService createSwaggerService(Set<OpenApiCustomizer> customizers) {
+        return new SwaggerService(createModels(specs, prepareCustomizers(customizers)));
     }
 
-    private Map<String, OpenApiModel> createModels(Map<String, OpenApiModelFactory> specs) {
+    private List<OpenApiCustomizer> prepareCustomizers(Set<OpenApiCustomizer> diCustomizers) {
+
+        // start with standard customizers, so that the user-provided ones can fix the model to
+        // their liking
+
+        List<OpenApiCustomizer> customizers = new ArrayList<>(2 + diCustomizers.size());
+        customizers.add(new PathSortingCustomizer());
+        customizers.add(new SchemasSortingCustomizer());
+        customizers.addAll(diCustomizers);
+
+        return customizers;
+    }
+
+    private Map<String, OpenApiModel> createModels(
+            Map<String, OpenApiModelFactory> specs,
+            List<OpenApiCustomizer> customizers) {
+
         Map<String, OpenApiModel> models = new HashMap<>();
-        if (specs == null) {
-            specs = new HashMap<>();
-        }
-        specs.values().stream()
-                .map(swaggerSpec -> swaggerSpec.createModel(overrideSpec, pretty))
-                .filter(Optional::isPresent)
-                .map(Optional::get)
-                .forEach(m -> indexByPath(models, m));
+
+        Map<String, OpenApiModelFactory> localSpecs = specs != null ? specs : new HashMap<>();
+        localSpecs.forEach((n, f) -> loadModel(models, n, f, customizers));
+
         return models;
+    }
+
+    private void loadModel(
+            Map<String, OpenApiModel> models,
+            String name,
+            OpenApiModelFactory factory,
+            List<OpenApiCustomizer> customizers) {
+        factory.createModel(name, overrideSpec, pretty, customizers).ifPresent(m -> indexByPath(models, m));
     }
 
     private void indexByPath(Map<String, OpenApiModel> models, OpenApiModel model) {